+++ spec/rubocop/cop/style/lambda_spec.rb
@@ -23,6 +23,13 @@ describe RuboCop::Cop::Style::Lambda, :config do
     end
   end
 
+  shared_examples 'does not auto-correct' do
+    it 'does not autocorrect' do
+      expect(autocorrect_source(cop, source)).to eq(source.join("\n"))
+      expect(cop.offenses.map(&:corrected?)).to eq [false]
+    end
+  end
+
   context 'with enforced `lambda` style' do
     let(:cop_config) { { 'EnforcedStyle' => 'lambda' } }
 
@@ -338,15 +345,21 @@ describe RuboCop::Cop::Style::Lambda, :config do
          'end, source: cats']
       end
 
-      it 'registers an offense' do
-        inspect_source(cop, source)
-        expect(cop.offenses.size).to eq 1
+      it_behaves_like 'registers an offense',
+                      'Use the `lambda` method for multiline lambdas.'
+      it_behaves_like 'does not auto-correct'
+    end
+
+    context 'with a multiline lambda literal as a keyword argument' do
+      let(:source) do
+        ['has_many opt: -> do',
+         '  where(cats: Cat.young.where_values_hash)',
+         'end']
       end
 
-      it 'does not auto-correct' do
-        expect(autocorrect_source(cop, source)).to eq(source.join("\n"))
-        expect(cop.offenses.map(&:corrected?)).to eq [false]
-      end
+      it_behaves_like 'registers an offense',
+                      'Use the `lambda` method for multiline lambdas.'
+      it_behaves_like 'does not auto-correct'
     end
   end
 end